// packages/codegen/test/fixtures/basic/normalizer.ts
// Auto-generated by @gqlens/codegen
import type { NormalizerEntry } from "@gqlens/core/codegen";

export const normalizerEntries: NormalizerEntry[] = [
  { type: "User", fields: [
    { responseKey: "id", cacheKey: "id" },
    { responseKey: "name", cacheKey: "name" },
    { responseKey: "avatar", cacheKey: "avatar" },
    { responseKey: "online", cacheKey: "online" },
    { responseKey: "posts", cacheKey: "posts", nestedType: "Post", isList: true },
  ] },
  { type: "Post", fields: [
    { responseKey: "id", cacheKey: "id" },
    { responseKey: "title", cacheKey: "title" },
    { responseKey: "content", cacheKey: "content" },
    { responseKey: "tags", cacheKey: "tags", isList: true },
    { responseKey: "author", cacheKey: "author", nestedType: "User" },
    { responseKey: "comments", cacheKey: "comments", nestedType: "Comment", isList: true },
  ] },
  { type: "Comment", fields: [
    { responseKey: "id", cacheKey: "id" },
    { responseKey: "body", cacheKey: "body" },
    { responseKey: "author", cacheKey: "author", nestedType: "User" },
  ] },
  { type: "Cat", fields: [
    { responseKey: "id", cacheKey: "id" },
    { responseKey: "name", cacheKey: "name" },
    { responseKey: "meows", cacheKey: "meows" },
  ] },
  { type: "Dog", fields: [
    { responseKey: "id", cacheKey: "id" },
    { responseKey: "name", cacheKey: "name" },
    { responseKey: "barks", cacheKey: "barks" },
  ] },
];